Just added some whitespace to get a commit message.
I just move around a ton of files to make this deployable to a remote server, so you'll probably want to commit everything you wanted to change before updating to this changesset.. You can now deploy this app to a CleverCloud server, hoorah! For some reason heroku wasn't playing nice.